requirements.txt
streamlit
pandas
matplotlib
gspread
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib